Make gometalinter happy by renaming remove _ from _spec packages
