--- web/apps_spec/empty.go
@@ -1,1 +1,0 @@
-package apps_func
