+++ web/appsspec/apps_test.go
@@ -1,7 +1,7 @@
 // apps_test package is introduced to avoid circular dependencies since this
 // particular test requires to depend on routing directly to expose the API
 // and the APP server.
-package apps_func
+package appsspec
 
 import (
 	"bytes"
