--- web/auth_spec/empty.go
@@ -1,1 +1,0 @@
-package auth_func
